requirements.txt
dearpygui==2.0.0
schedule==1.2.2
pyserial==3.5
